+++ src/Readarr.Api.V1/BookFiles/BookFileResource.cs
@@ -22,6 +22,7 @@ namespace Readarr.Api.V1.BookFiles
         public BookFileMediaType MediaType { get; set; }
         public string ConversionError { get; set; }
         public bool ConversionErrorIsDrm { get; set; }
+        public int PageCount { get; set; }
 
         public bool QualityCutoffNotMet { get; set; }
         public ParsedTrackInfo AudioTags { get; set; }
@@ -62,6 +63,7 @@ namespace Readarr.Api.V1.BookFiles
                 MediaType = model.MediaType != BookFileMediaType.Unknown
                     ? model.MediaType
                     : MediaFileExtensions.GetMediaTypeForPath(model.Path),
+                PageCount = model.Edition.Value?.PageCount ?? 0,
                 ConversionError = model.ConversionError,
                 ConversionErrorIsDrm = model.ConversionErrorIsDrm
             };
@@ -89,6 +91,7 @@ namespace Readarr.Api.V1.BookFiles
                 MediaType = model.MediaType != BookFileMediaType.Unknown
                     ? model.MediaType
                     : MediaFileExtensions.GetMediaTypeForPath(model.Path),
+                PageCount = model.Edition.Value?.PageCount ?? 0,
                 QualityCutoffNotMet = upgradableSpecification.QualityCutoffNotMet(author.QualityProfile.Value, model.Quality),
                 IndexerFlags = (int)model.IndexerFlags,
                 ConversionError = model.ConversionError,